typescript definitons done, all tests pass, removing javascript tests
